Use longer names instead of mod for alignment calculators.
This works around rust-lang/rust#83583 by avoiding generating a mod block. This
code previously worked everywhere, but tripped this deprecation lint if
Crevice's derives were used on a struct inside a function.

In order to have consistent name resolution, we instead blast all the generated
functions into the main scope. This is... not great, but there is no other way
to let code inside a mod inside a fn refer to other items inside the fn.